Add files via upload
#Data set: 
The dataset was obtained from the Kaggle website on the population of various countries. The dataset includes information on individual countries' names and regions, happiness scores and rankings, economic and health conditions, and so on.  The project's main objective is to get comparisons between different information and predict happiness scores using machine learning algorithms.
https://www.kaggle.com/datasets/sanjanchaudhari/population-dataset
The data set consists of 12 columns and 158 rows. 


#Data Preprocessing:
In the data preprocessing the null values, data types, and counts are checked. There are no duplicates or null values. All other indicators are float data types without integer happiness score, country, and region object types.

#Explore Data Analysis: 
Some insights were obtained from the analysis which gives an idea of the overall dataset. There are 158 different countries under 10 different regions. The mean happiness score is visualized by a barplot which groups by regions and plots economy, trust, freedom, and generosity with happiness score. The correlation score between different variables is shown by sns heatmap.


#Machine Learning Process: 
In the machine learning process of prediction, we don’t need the country name, region, and happiness rank. Out of these, all are numeric, especially float data types. So we don’t need conversion of categorical data into a numerical format and also scaling as all data are in a small range. The dataset was divided into dependent and independent variables. The train test split method was applied for training the model and after that, the accuracy of the model was checked by a test dataset which was 20% data from the whole dataset. The score of the Linear Regression model is 99.99% which is a good estimator.